math: Fix UB in dbl-64 lrint
UBSAN: Undefined behaviour in ../sysdeps/ieee754/dbl-64/s_lrint.c:99:30 left shift of 1048576 by 11 cannot be represented in type 'long int'

+++ sysdeps/ieee754/dbl-64/s_lrint.c
@@ -95,7 +95,7 @@ __lrint (double x)
 	  if (j0 == 20)
 	    result = (long int) i0;
 	  else
-	    result = ((long int) i0 << (j0 - 20)) | (i1 >> (52 - j0));
+	    result = (i0 << (j0 - 20)) | (i1 >> (52 - j0));
 	}
     }
   else
